Rename upt.txt to utp.txt
Five months ago (as shown by the pull request date), someone uploaded this file with the university's name, but it seems they made a mistake in the filename. Instead of UTP (Universidad Tecnológica del Perú / Technological University of Peru), they wrote UPT, which causes authentication errors for students because the system does not correctly recognize the university's domain (which should be UTP).

Could this be corrected to avoid further authentication issues for students using UTP email domains?